fact_checker/requirements.txt
langchain
langchain-openai
langchain-anthropic
langchain-google-genai
langchain-ollama
openai
anthropic
google-generativeai
ollama
pydantic>=2.0.0
python-dotenv
requests

